+++ zap/lib/src/http/interface/http_client_response.dart
@@ -1,10 +1,12 @@
+import 'package:zap/src/definitions.dart';
+
 import 'http_headers.dart';
 import 'redirect_info.dart';
 
 /// This interface is used to represent a response from a server.
 /// 
-/// It extends [Stream<List<int>>] to provide a stream of bytes which is the response body.
-abstract interface class HttpClientResponse implements Stream<List<int>> {
+/// It extends [BodyBytes] to provide a stream of bytes which is the response body.
+abstract interface class HttpClientResponse implements BodyBytes {
   /// Returns the status code.
   ///
   /// The status code must be set before the body is written